Hopefully useful player and hitbox class.
I ended up making an updated version of the player class which I'm hoping might also act as a small template for the enemy class (since it should pretty much be the same but just without the player input functions).
If the following html is added to the body of our index.html it should work fine when we call the player constructor in our main game script.
<p id = "javascriptcode">JS Code</p>
<script type="text/javascript" src="player_and_hitbox.js"></script>

What I thought this could really help us with is so that when we do start building the room/level at the start of each game at least we have instant control of where the player (and maybe even the enemies if we recycle the player class) will spawn since I've added that to the constructor. 

It doesn't run fully just yet. We need to add calls to the player's updatePosition() function in the main game loop and hopefully that should sync it nicely with the other things we've currently got. I've not made it a subclass of the other game objects yet so that still needs to be done. Furthermore collision detection between the player and hitboxes or other players still needs to be implemented.

I've made spacebar key the 'attack' button which should briefly create a new hitbox object. The coordinates of the hitbox will be relative to the player's current position and will also depend on which direction the player is facing (determined by the player's getPlayerState() function). The Hitbox class still needs some work though (areas such as lifespan and possibly giving it a sprite).

That's all for this commit, hopefully some stuff will be of use. :)