Started code for player collision detection
I tried adding some code for the player's collision detection. It will need more work:
Currently, object variable is just a place holder for the gameobject that the player's position will be compared to. An idea might be to create an array of tiles/blocks that are solid which we could then itterate through during the collision code to check if the player has collided with anything that should stop it.