Changed login/stats button to change callback rather than use 2 buttons
Also removed player from page when they sign out, this should also stop
the clock but there is an issue with scope and the stop function cannot
be accessed at this point. I think that when the user signs out they
should be denied access to the start loop button as well.