Changed guest code so their details are no longer stored in db
Also changed the statistics/sign out window so it is only reliant on
getting the db info at when the player logs in. It then uses the
socket.player info to update the stats based on current scores.

Once we have scoring/kills/wins implemented call the
updateUserStatVals() function to update the stat window

Logging in with guest for multiplayer is still broken