Synced timer
Still needs to have a proper way to start it, other than when server
starts running.
There is now a reset, this takes 5 secs but it counts down into the
negatives while waiting (this should be changed).
Pretty sure this emits to all clients regardless of what level they are
in, this should probably be changed (this will be easier once the match
start thing is implemented)