Added in support for 3 spawn points for each team
Changed level editor a bit so spawn points are to be entered in the
following format:

(2.5 -2.5),(9.5 -7.5),(6.5 -5.5),(45.5 -29.5),(38.5 -24.5),(41.5 -26.5)

There must be 6 spawn points, encapsulated in brackets with spaces to
separate x and y coords within each point and the spawn points should be
separated by commas (no spaces)
This shouldn't be a problem as teams are 3 vs 3 so the first 3 pos are
for blue team spawn and the last 3 are for red team.

Changed the map to a new one, can modify if needed. This one still needs
to have the actual walls put in (didn't remember how to when I done it).
Will do that soon though.