Added in bases to the world
To add in a base in the level editor, the base tiles are separate tiles.
At the min they only consist of dark and light wood tiles for both blue
and red team. Don't really need more than them at this point but for
future reference if you add another one in using pixilart, set the
opacity of the colour to 30% to make it match up and add in the tile Id
to the if statements in setTileAt in editor.js

Currently playerScript is used to detect whether a player of the correct
team is standing on their base. This should be adapted in the future
once we have player-ball interaction sorted to add a point and change
image back to default