Added initial support for session stat updating in the stat window
The stat window updates when the activation button is clicked on, still
need to wipe game stats for a logged in user on match reset.
If the user is a guest their session stats will be updated with the game
stats and then the game stats will be wiped, this is so showing game
stats in the leaderboard will be possible. This has still to be
implemented.

Now should go onto updating the database with the game stats at match
end for logged in users