Removed levels[0] as default level and added in level deletion on empty
- Server now checks to see if there are any current levels, if not it
creates one with id 1. This will only be the case after a server restart

- Server now checks the player count at the end of a match. If the
player count is 0, that level will be deleted. It doesn't decrement
levelCount as this could cause some issues if all players from level 1
leave while there are some players in level 2 left - don't want to
levels assigned to the same id